Removing AI guesses at object names, import cleanup, moving AuthContext.tsx to /contexts/ directory
